fix *_waiting list usage, service loss no longer permanent
The lists *_waiting now contain topics, services or actions which we are
expecting, but do not currently exist. Once it comes into existence, we remove
it from this list.

When services disconnect, their loss is no longer permanent. This had to do with
the services being removed and not added to the waiting list.

Fixes issue #21.